backend/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
pydantic>=2.5

groq==0.9.0
httpx==0.27.0
firebase-admin
python-dotenv

